requirements.txt
click>=8.0
pyyaml>=6.0
openai>=1.0
anthropic>=0.18
aiohttp>=3.9
gitpython>=3.1
tree-sitter>=0.20
rich>=13.0
